updated: visualizations to use float instead of signed 16bit

+++ xbmc/visualizations/Milkdrop/MilkdropXBMC.cpp
@@ -117,7 +117,7 @@ unsigned char waves[2][576];
 //-- Audiodata ----------------------------------------------------------------
 // Called by XBMC to pass new audio data to the vis
 //-----------------------------------------------------------------------------
-extern "C" void AudioData(const short* pAudioData, int iAudioDataLength, float *pFreqData, int iFreqDataLength)
+extern "C" void AudioData(const float* pAudioData, int iAudioDataLength, float *pFreqData, int iFreqDataLength)
 {
 	int ipos=0;
 	while (ipos < 576)
